split up month-data to working and holidays
