+++ src/lib/utils.js
@@ -1,0 +1,5 @@
+
+export function getDaysInMonth (year, month) {
+  // first day in month is 1. 0 is the one before --> last day of month!
+  return new Date(year, month + 1, 0).getDate()
+}
